pyproject.toml
[project]
name = "groupup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pwdpy>=1.0.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.48",
]

[tool.fastapi]
entrypoint = "src.main:app"
